implemented Host-owned NetIds for mobsync
