+++ FakeDeath/FakeDeath.cs
@@ -6,7 +6,6 @@ using dc.en;
 using dc.tool.atk;
 using dc.tool.mainSkills;
 using dc.ui;
-using dc.cine;
 using DeadCellsMultiplayerMod.Ghost.GhostBase;
 using DeadCellsMultiplayerMod.MultiplayerModUI.lifeUI;
 using ModCore.Modules;
@@ -21,76 +20,11 @@ namespace DeadCellsMultiplayerMod
         private bool _allDownedRestartQueued;
         private long _allDownedRestartAtTicks;
         private const double AllDownedGameOverDelaySeconds = 0.35;
-        // Cursed/environmental death hooks can run from inside attack/cooldown iteration. Constructing
-        // DeadBase or GameOver re-entrantly from that stack can leave HashLink in a permanent main-loop
-        // stall. Defer corpse-cinematic creation until a later normal frame and serialize flow updates.
-        private const double LocalDeadCineCreateDelaySeconds = 0.12;
-        private long _localDeadCineCreateAfterTicks;
         private bool _fakeDeathFlowInProgress;
-        private bool _hasLocalDownedAnchor;
-        private double _localDownedAnchorX;
-        private double _localDownedAnchorY;
-        private const double DownedCorpseMaxDriftPx = 96.0;
-        private const double DownedCorpseMaxDriftSq = DownedCorpseMaxDriftPx * DownedCorpseMaxDriftPx;
-        // Keep enough grounded history to recover a corpse even after a long pit fall or a
-        // several-second trap sequence. The old 18-sample buffer only covered about two seconds,
-        // which was too short for many off-map deaths.
-        private const int SafeReviveAnchorHistorySize = 96;
-        private const double SafeReviveAnchorSampleSeconds = 0.12;
-        private const double SafeReviveAnchorPreferredAgeSeconds = 0.75;
-        private const double SafeReviveAnchorMaxAgeSeconds = 12.0;
-        private const double EnvironmentalDamageContextSeconds = 2.5;
-        // A sample recorded immediately before or during contact with spikes/lava is not safe.
-        // Quarantine new samples briefly after environmental damage and remove the recent tail.
-        private const double SafeAnchorEnvironmentalQuarantineSeconds = 2.5;
-        private const double SafeAnchorEnvironmentalPurgeSeconds = 2.25;
-        // Downed bodies should never remain at the lethal coordinate. Prefer an older point far
-        // enough away to be outside even a wide spike strip, then fall back to the oldest valid
-        // same-room anchor rather than the death position.
-        private const double HazardRecoveryPreferredAgeSeconds = 1.25;
-        private const double HazardRecoveryFallbackAgeSeconds = 2.25;
-        private const double HazardRecoveryMinDistancePx = 144.0;
-        private const double HazardRecoveryMinDistanceSq = HazardRecoveryMinDistancePx * HazardRecoveryMinDistancePx;
-
-        private readonly struct SafeReviveAnchorSample
-        {
-            public readonly double X;
-            public readonly double Y;
-            public readonly long Ticks;
-            public readonly string LevelId;
-
-            public SafeReviveAnchorSample(double x, double y, long ticks, string levelId)
-            {
-                X = x;
-                Y = y;
-                Ticks = ticks;
-                LevelId = levelId ?? string.Empty;
-            }
-        }
-
-        private readonly SafeReviveAnchorSample[] _safeReviveAnchors = new SafeReviveAnchorSample[SafeReviveAnchorHistorySize];
-        private int _safeReviveAnchorCount;
-        private int _safeReviveAnchorWriteIndex;
-        private long _nextSafeReviveAnchorSampleTicks;
-        private string _safeReviveAnchorLevelId = string.Empty;
-        private bool _hasSafeAnchorMotionProbe;
-        private double _lastSafeAnchorMotionProbeY;
-        private string _safeAnchorMotionProbeLevelId = string.Empty;
-        private bool _localDownedUsesRecoveryAnchor;
         private bool _localDownedGravityCaptured;
         private bool _localDownedOriginalHasGravity = true;
-        private bool _lastLocalDamageWasEnvironmental;
-        private long _lastLocalDamageContextTicks;
         private readonly HashSet<int> _scratchRemoteActiveIds = new();
-        private readonly HashSet<int> _scratchActiveCorpseIds = new();
         private readonly List<int> _scratchStaleRemoteIds = new();
-        private readonly List<int> _scratchStaleCorpseIds = new();
-        // Environmental deaths can temporarily publish an invalid room marker or put the remote
-        // GhostKing into the engine's out-of-game state. Keep a brief revive grace period so the
-        // normal snapshot stream can reattach the same remote shell without requiring a sublevel
-        // round-trip to rebuild it.
-        private readonly Dictionary<int, long> _remoteReviveVisibilityGraceUntilTicks = new();
-        private const double RemoteReviveVisibilityGraceSeconds = 3.0;
 
         private void Hook_Hero_onHeroDie(Hook_Hero.orig_onHeroDie orig, Hero self)
         {
@@ -265,505 +199,6 @@ namespace DeadCellsMultiplayerMod
             // map from a death hook risks a delayed HashLink null/type crash in the main loop.
         }
 
-        private void RecordLocalDamageContext(AttackData? attack)
-        {
-            var now = Stopwatch.GetTimestamp();
-            _lastLocalDamageContextTicks = now;
-            _lastLocalDamageWasEnvironmental = IsEnvironmentalDamageSource(attack);
-            if (_lastLocalDamageWasEnvironmental)
-                InvalidateRecentSafeReviveAnchors(now, SafeAnchorEnvironmentalPurgeSeconds);
-        }
-
-        private void InvalidateRecentSafeReviveAnchors(long now, double seconds)
-        {
-            if (_safeReviveAnchorCount <= 0 || seconds <= 0.0)
-                return;
-
-            var cutoffTicks = now - (long)(Stopwatch.Frequency * seconds);
-            for (var i = 0; i < _safeReviveAnchors.Length; i++)
-            {
-                var sample = _safeReviveAnchors[i];
-                if (sample.Ticks > 0 && sample.Ticks >= cutoffTicks)
-                    _safeReviveAnchors[i] = default;
-            }
-        }
-
-        private static bool IsEnvironmentalDamageSource(AttackData? attack)
-        {
-            if (attack == null)
-                return true;
-
-            dc.Entity? source = null;
-            try { source = attack.source; } catch { }
-            if (source == null)
-                return true;
-            if (source is Mob || source is Hero || source is GhostKing)
-                return false;
-
-            try
-            {
-                var name = source.GetType().Name ?? string.Empty;
-                if (name.IndexOf("Bullet", StringComparison.OrdinalIgnoreCase) >= 0 ||
-                    name.IndexOf("Projectile", StringComparison.OrdinalIgnoreCase) >= 0 ||
-                    name.IndexOf("Arrow", StringComparison.OrdinalIgnoreCase) >= 0 ||
-                    name.IndexOf("Shot", StringComparison.OrdinalIgnoreCase) >= 0)
-                {
-                    return false;
-                }
-            }
-            catch
-            {
-            }
-
-            return true;
-        }
-
-        private void UpdateLocalSafeReviveAnchor(Hero? hero)
-        {
-            if (hero == null || _localFakeDead || _localDeathConversionInProgress)
-                return;
-
-            try
-            {
-                if (hero.destroyed || hero.life <= 0 || hero._level == null || hero.spr == null)
-                    return;
-                if (hero.isOutOfGame)
-                    return;
-            }
-            catch
-            {
-                return;
-            }
-
-            var level = GetCurrentLevelId();
-            if (string.IsNullOrWhiteSpace(level))
-                return;
-            if (!string.Equals(_safeReviveAnchorLevelId, level, StringComparison.Ordinal))
-                ResetSafeReviveAnchorHistory(level);
-
-            var now = Stopwatch.GetTimestamp();
-            if (_lastLocalDamageWasEnvironmental && _lastLocalDamageContextTicks > 0 &&
-                now - _lastLocalDamageContextTicks <
-                (long)(Stopwatch.Frequency * SafeAnchorEnvironmentalQuarantineSeconds))
-            {
-                return;
-            }
-
-            if (_nextSafeReviveAnchorSampleTicks != 0 && now < _nextSafeReviveAnchorSampleTicks)
-                return;
-            _nextSafeReviveAnchorSampleTicks = now +
-                (long)(Stopwatch.Frequency * SafeReviveAnchorSampleSeconds);
-
-            try
-            {
-                var verticalMotion = Math.Abs(hero.dy) + Math.Abs(hero.bdy);
-                if (!double.IsFinite(verticalMotion) || verticalMotion > 0.45)
-                    return;
-            }
-            catch
-            {
-            }
-
-            double x;
-            double y;
-            if (!TryGetHeroLogicalPixelPosition(hero, out x, out y))
-            {
-                try
-                {
-                    x = hero.get_targetSprPosX();
-                    y = hero.get_targetSprPosY();
-                }
-                catch
-                {
-                    try
-                    {
-                        x = hero.spr?.x ?? 0.0;
-                        y = hero.spr?.y ?? 0.0;
-                    }
-                    catch
-                    {
-                        return;
-                    }
-                }
-            }
-
-            if (!TryProjectHeroPositionToSafeGround(hero, x, y, out var safeX, out var safeY))
-                return;
-
-            // Do not certify positions while a platform/elevator is carrying the hero vertically.
-            // Hero.dy can remain near zero on moving platforms, so compare logical floor Y across
-            // samples as a second independent stability check. The first stable probe after any
-            // vertical movement is only observed; the following stable probe becomes an anchor.
-            if (!_hasSafeAnchorMotionProbe ||
-                !string.Equals(_safeAnchorMotionProbeLevelId, level, StringComparison.Ordinal))
-            {
-                _hasSafeAnchorMotionProbe = true;
-                _lastSafeAnchorMotionProbeY = safeY;
-                _safeAnchorMotionProbeLevelId = level;
-                return;
-            }
-
-            var floorDeltaY = Math.Abs(safeY - _lastSafeAnchorMotionProbeY);
-            _lastSafeAnchorMotionProbeY = safeY;
-            if (!double.IsFinite(floorDeltaY) || floorDeltaY > 2.0)
-                return;
-
-            _safeReviveAnchors[_safeReviveAnchorWriteIndex] =
-                new SafeReviveAnchorSample(safeX, safeY, now, level);
-            _safeReviveAnchorWriteIndex = (_safeReviveAnchorWriteIndex + 1) % SafeReviveAnchorHistorySize;
-            if (_safeReviveAnchorCount < SafeReviveAnchorHistorySize)
-                _safeReviveAnchorCount++;
-        }
-
-        private void ResetSafeReviveAnchorHistory(string levelId)
-        {
-            Array.Clear(_safeReviveAnchors, 0, _safeReviveAnchors.Length);
-            _safeReviveAnchorCount = 0;
-            _safeReviveAnchorWriteIndex = 0;
-            _nextSafeReviveAnchorSampleTicks = 0;
-            _safeReviveAnchorLevelId = levelId ?? string.Empty;
-            _hasSafeAnchorMotionProbe = false;
-            _lastSafeAnchorMotionProbeY = 0.0;
-            _safeAnchorMotionProbeLevelId = levelId ?? string.Empty;
-        }
-
-        private bool TryGetSafeReviveAnchor(string levelId, long now, bool preferOlderSample, out double x, out double y)
-        {
-            x = 0.0;
-            y = 0.0;
-            if (_safeReviveAnchorCount <= 0 || string.IsNullOrWhiteSpace(levelId))
-                return false;
-
-            SafeReviveAnchorSample? newestValid = null;
-            for (var i = 0; i < _safeReviveAnchorCount; i++)
-            {
-                var index = (_safeReviveAnchorWriteIndex - 1 - i + SafeReviveAnchorHistorySize) %
-                            SafeReviveAnchorHistorySize;
-                var sample = _safeReviveAnchors[index];
-                if (sample.Ticks <= 0 ||
-                    !string.Equals(sample.LevelId, levelId, StringComparison.Ordinal))
-                {
-                    continue;
-                }
-
-                var ageSeconds = (now - sample.Ticks) / (double)Stopwatch.Frequency;
-                if (ageSeconds < 0.0 || ageSeconds > SafeReviveAnchorMaxAgeSeconds)
-                    continue;
-
-                newestValid ??= sample;
-                if (!preferOlderSample || ageSeconds >= SafeReviveAnchorPreferredAgeSeconds)
-                {
-                    x = sample.X;
-                    y = sample.Y;
-                    return true;
-                }
-            }
-
-            if (newestValid.HasValue)
-            {
-                x = newestValid.Value.X;
-                y = newestValid.Value.Y;
-                return true;
-            }
-
-            return false;
-        }
-
-        private bool TryGetHazardRecoveryAnchor(
-            string levelId,
-            long now,
-            double deathX,
-            double deathY,
-            out double x,
-            out double y)
-        {
-            x = 0.0;
-            y = 0.0;
-            if (_safeReviveAnchorCount <= 0 || string.IsNullOrWhiteSpace(levelId))
-                return false;
-
-            SafeReviveAnchorSample? agedFallback = null;
-            SafeReviveAnchorSample? oldestValid = null;
-            var hasFiniteDeathPosition = double.IsFinite(deathX) && double.IsFinite(deathY);
-
-            // Search from newest to oldest. Prefer a point that is both old enough to pre-date the
-            // hazard contact and far enough away that it is unlikely to still be inside the same
-            // spike bed, pit edge, lava strip, or trap volume.
-            for (var i = 0; i < _safeReviveAnchorCount; i++)
-            {
-                var index = (_safeReviveAnchorWriteIndex - 1 - i + SafeReviveAnchorHistorySize) %
-                            SafeReviveAnchorHistorySize;
-                var sample = _safeReviveAnchors[index];
-                if (sample.Ticks <= 0 ||
-                    !string.Equals(sample.LevelId, levelId, StringComparison.Ordinal))
-                {
-                    continue;
-                }
-
-                var ageSeconds = (now - sample.Ticks) / (double)Stopwatch.Frequency;
-                if (ageSeconds < 0.0 || ageSeconds > SafeReviveAnchorMaxAgeSeconds)
-                    continue;
-
-                oldestValid = sample;
-                if (!agedFallback.HasValue && ageSeconds >= HazardRecoveryFallbackAgeSeconds)
-                    agedFallback = sample;
-
-                var separatedFromDeath = true;
-                if (hasFiniteDeathPosition)
-                {
-                    var dx = sample.X - deathX;
-                    var dy = sample.Y - deathY;
-                    separatedFromDeath = dx * dx + dy * dy >= HazardRecoveryMinDistanceSq;
-                }
-
-                if (ageSeconds >= HazardRecoveryPreferredAgeSeconds && separatedFromDeath)
-                {
-                    x = sample.X;
-                    y = sample.Y;
-                    return true;
-                }
-            }
-
-            // A player can die while almost stationary on a trap, so distance may not produce a
-            // candidate. In that case use an older grounded sample rather than leaving the corpse
-            // in the hazard. The final fallback is the oldest still-valid same-room sample.
-            var fallback = oldestValid ?? agedFallback;
-            if (!fallback.HasValue)
-                return false;
-
-            x = fallback.Value.X;
-            y = fallback.Value.Y;
-            return true;
-        }
-
-        private bool TryGetLivingTeammateSafeAnchor(Hero hero, out double x, out double y)
-        {
-            x = 0.0;
-            y = 0.0;
-            if (hero == null)
-                return false;
-
-            var net = _net;
-            var localId = net?.id ?? 0;
-            for (var i = 0; i < clients.Length; i++)
-            {
-                var client = clients[i];
-                if (client == null)
-                    continue;
-
-                try
-                {
-                    if (client.destroyed || client._level == null || client.spr == null)
-                        continue;
-                }
-                catch
-                {
-                    continue;
-                }
-
-                var remoteId = clientIds[i];
-                if (remoteId <= 0 || (localId > 0 && remoteId == localId) || IsRemotePlayerDowned(remoteId))
-                    continue;
-
-                double remoteX;
-                double remoteY;
-                if (!TryGetGhostLogicalPixelPosition(client, out remoteX, out remoteY))
-                {
-                    try
-                    {
-                        remoteX = client.get_targetSprPosX();
-                        remoteY = client.get_targetSprPosY();
-                    }
-                    catch
-                    {
-                        try
-                        {
-                            remoteX = client.spr?.x ?? 0.0;
-                            remoteY = client.spr?.y ?? 0.0;
-                        }
-                        catch
-                        {
-                            continue;
-                        }
-                    }
-                }
-
-                if (double.IsFinite(remoteX) && double.IsFinite(remoteY))
-                {
-                    x = remoteX;
-                    y = remoteY;
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        private static bool TryGetHeroLogicalPixelPosition(Hero hero, out double x, out double y)
-        {
-            x = 0.0;
-            y = 0.0;
-            if (hero == null)
-                return false;
-
-            try
-            {
-                x = (hero.cx + hero.xr) * 24.0;
-                y = (hero.cy + hero.yr) * 24.0;
-                return double.IsFinite(x) && double.IsFinite(y);
-            }
-            catch
-            {
-                x = 0.0;
-                y = 0.0;
-                return false;
-            }
-        }
-
-        private static bool TryGetGhostLogicalPixelPosition(GhostKing king, out double x, out double y)
-        {
-            x = 0.0;
-            y = 0.0;
-            if (king == null)
-                return false;
-
-            try
-            {
-                x = (king.cx + king.xr) * 24.0;
-                y = (king.cy + king.yr) * 24.0;
-                return double.IsFinite(x) && double.IsFinite(y);
-            }
-            catch
-            {
-                x = 0.0;
-                y = 0.0;
-                return false;
-            }
-        }
-
-        // Do not call LevelMap.getGroundYr from the managed per-frame update path. On the current
-        // DCCM/GameProxy combination that native bridge can receive the wrong HashLink receiver and
-        // terminate the game with "Can't cast tool.CPoint to level.LevelMap". Safe anchors are
-        // therefore selected only from finite, alive, same-level positions with low vertical motion.
-        // The history itself provides the ground/reachability guarantee without touching LevelMap.
-        private static bool TryProjectHeroPositionToSafeGround(Hero hero, double x, double y, out double safeX, out double safeY)
-        {
-            safeX = x;
-            safeY = y;
-            if (hero == null || !double.IsFinite(x) || !double.IsFinite(y))
-                return false;
-
-            try
-            {
-                if (hero.destroyed || hero._level == null || hero.spr == null || hero.isOutOfGame)
-                    return false;
-
-                var verticalMotion = Math.Abs(hero.dy) + Math.Abs(hero.bdy);
-                if (!double.IsFinite(verticalMotion) || verticalMotion > 0.45)
-                    return false;
-            }
-            catch
-            {
-                return false;
-            }
-
-            return true;
-        }
-
-        private bool IsUnsafeLocalDeathPosition(Hero hero, double x, double y)
-        {
-            if (hero == null || !double.IsFinite(x) || !double.IsFinite(y))
-                return true;
-
-            try
-            {
-                if (hero.isOutOfGame)
-                    return true;
-            }
-            catch
-            {
-            }
-
-            try
-            {
-                var verticalMotion = Math.Abs(hero.dy) + Math.Abs(hero.bdy);
-                if (double.IsFinite(verticalMotion) && verticalMotion > 0.65)
-                    return true;
-            }
-            catch
-            {
-            }
-
-            // Finite, in-game positions with low vertical motion are usable. Environmental
-            // damage still selects an older history sample in ResolveLocalDownedAnchor.
-            return false;
-        }
-
-        private void ResolveLocalDownedAnchor(Hero hero, double deathX, double deathY, out double downedX, out double downedY)
-        {
-            downedX = deathX;
-            downedY = deathY;
-            _localDownedUsesRecoveryAnchor = false;
-
-            var now = Stopwatch.GetTimestamp();
-            var level = GetCurrentLevelId();
-            var recentEnvironmentalDamage = _lastLocalDamageWasEnvironmental &&
-                _lastLocalDamageContextTicks > 0 &&
-                now - _lastLocalDamageContextTicks <=
-                (long)(Stopwatch.Frequency * EnvironmentalDamageContextSeconds);
-            var unsafePosition = IsUnsafeLocalDeathPosition(hero, deathX, deathY);
-
-            // Always prefer a confirmed earlier anchor. This is deliberately not limited to
-            // deaths that were correctly classified as environmental: some spike/pit kill paths
-            // bypass onDamage and only reach Hero.kill/onDie, which previously left the body at
-            // the lethal coordinate. The history selection requires age and separation first,
-            // then uses the oldest valid same-room sample as a guaranteed reachable fallback.
-            if (TryGetHazardRecoveryAnchor(level, now, deathX, deathY, out var safeX, out var safeY))
-            {
-                downedX = safeX;
-                downedY = safeY;
-                _localDownedUsesRecoveryAnchor = true;
-                Logger.Information(
-                    "[NetMod][ReviveAnchor] selected prior safe downed anchor environmental={Environmental} unsafe={Unsafe} deathX={DeathX:0.0} deathY={DeathY:0.0} safeX={SafeX:0.0} safeY={SafeY:0.0}",
-                    recentEnvironmentalDamage,
-                    unsafePosition,
-                    deathX,
-                    deathY,
-                    downedX,
-                    downedY);
-            }
-            else if (TryGetLivingTeammateSafeAnchor(hero, out safeX, out safeY))
-            {
-                // Very early room deaths can occur before the local history contains a sample.
-                // A living teammate is then the only known reachable in-room location.
-                downedX = safeX;
-                downedY = safeY;
-                _localDownedUsesRecoveryAnchor = true;
-                Logger.Information(
-                    "[NetMod][ReviveAnchor] used living teammate fallback safeX={SafeX:0.0} safeY={SafeY:0.0}",
-                    downedX,
-                    downedY);
-            }
-            else if (!recentEnvironmentalDamage &&
-                     !unsafePosition &&
-                     TryProjectHeroPositionToSafeGround(hero, deathX, deathY, out var groundX, out var groundY))
-            {
-                // Last-resort only: no history and no living teammate. Keep a normal non-hazard
-                // death where it occurred rather than manufacturing an unverified coordinate.
-                downedX = groundX;
-                downedY = groundY;
-            }
-            else if (TryGetSafeReviveAnchor(level, now, preferOlderSample: true, out safeX, out safeY))
-            {
-                downedX = safeX;
-                downedY = safeY;
-                _localDownedUsesRecoveryAnchor = true;
-            }
-
-            _lastLocalDamageWasEnvironmental = false;
-            _lastLocalDamageContextTicks = 0;
-        }
-
         private bool CanConvertLocalHeroDeathToFakeDeath(Hero? self, NetNode? net)
         {
             if (_netRole == NetRole.None || net == null || !net.IsAlive)
@@ -811,97 +246,6 @@ namespace DeadCellsMultiplayerMod
             }
 
             return true;
-        }
-
-        private static bool IsVanillaHeroDeathCineActive()
-        {
-            try
-            {
-                var cine = dc.pr.Game.Class.ME?.curCine;
-                return cine is HeroDeath ||
-                       cine is HeroDeathBase ||
-                       cine is HeroDeathContinue ||
-                       cine is HeroDeathRespawn ||
-                       cine is HeroDeathDLCP;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-
-        private bool ShouldSuppressVanillaHeroDeathCinematic(Hero? lostBody)
-        {
-            return _netRole != NetRole.None &&
-                   _net != null &&
-                   _net.IsAlive &&
-                   me != null &&
-                   lostBody != null &&
-                   ReferenceEquals(lostBody, me);
-        }
-
-        private bool SuppressVanillaHeroDeathCinematic(Hero? lostBody, dc.GameCinematic? cine)
-        {
-            if (!ShouldSuppressVanillaHeroDeathCinematic(lostBody))
-                return false;
-
-            if (!_localFakeDead && lostBody != null && _net != null)
-                EnterLocalFakeDeath(lostBody, _net);
-
-            try
-            {
-                var game = dc.pr.Game.Class.ME;
-                if (game != null && cine != null && ReferenceEquals(game.curCine, cine))
-                    game.curCine = null;
-            }
-            catch
-            {
-            }
-
-            // Constructor hooks run before the vanilla cinematic object is initialized. Calling
-            // destroy/disposeImmediately on that half-constructed object can tear down unrelated
-            // hero state. Simply skip the constructor; startDeathCine/kill are already redirected.
-            return true;
-        }
-
-        private void Hook__HeroDeath__constructor__(Hook__HeroDeath.orig___constructor__ orig, HeroDeath e, Hero lostBody, bool fromMob)
-        {
-            if (SuppressVanillaHeroDeathCinematic(lostBody, e))
-                return;
-
-            orig(e, lostBody, fromMob);
-        }
-
-        private void Hook__HeroDeathBase__constructor__(Hook__HeroDeathBase.orig___constructor__ orig, HeroDeathBase e, Hero lostBody, bool mob)
-        {
-            if (SuppressVanillaHeroDeathCinematic(lostBody, e))
-                return;
-
-            orig(e, lostBody, mob);
-        }
-
-        private void Hook__HeroDeathContinue__constructor__(Hook__HeroDeathContinue.orig___constructor__ orig, HeroDeathContinue e, Hero lostBody, bool keepBody)
-        {
-            if (SuppressVanillaHeroDeathCinematic(lostBody, e))
-                return;
-
-            orig(e, lostBody, keepBody);
-        }
-
-        private void Hook__HeroDeathRespawn__constructor__(Hook__HeroDeathRespawn.orig___constructor__ orig, HeroDeathRespawn e, Hero lostBody)
-        {
-            if (SuppressVanillaHeroDeathCinematic(lostBody, e))
-                return;
-
-            orig(e, lostBody);
-        }
-
-        private void Hook__HeroDeathDLCP__constructor__(Hook__HeroDeathDLCP.orig___constructor__ orig, HeroDeathDLCP e, Hero lostBody, bool fromMob)
-        {
-            if (SuppressVanillaHeroDeathCinematic(lostBody, e))
-                return;
-
-            orig(e, lostBody, fromMob);
         }
 
         private void Hook_Hero_startDeathCine(Hook_Hero.orig_startDeathCine orig, Hero self)
@@ -1237,284 +581,6 @@ namespace DeadCellsMultiplayerMod
 
             if (combatStateChanged)
                 global::DeadCellsMultiplayerMod.Mobs.MobsSynchronization.MobsSynchronization.NotifyPlayerCombatStateChanged("remote-down-pruned");
-        }
-
-        private void ApplyRemoteDownedGhostPositions(NetNode net)
-        {
-            if (net == null)
-                return;
-
-            if (_remoteDowned.Count == 0)
-            {
-                DisposeAllRemoteDownedCines();
-                for (int i = 0; i < clients.Length; i++)
-                {
-                    var client = clients[i];
-                    if (client != null)
-                    {
-                        try { client._targetable = true; } catch { }
-                    }
-                }
-                return;
-            }
-
-            var localId = net.id;
-            var localLevelId = GetCurrentLevelId();
-            _scratchActiveCorpseIds.Clear();
-            foreach (var state in _remoteDowned.Values)
-            {
-                if (state == null || state.UserId <= 0)
-                    continue;
-                if (!TryGetClientIndex(localId, state.UserId, out var index))
-                {
-                    DisposeRemoteDownedCine(state.UserId);
-                    continue;
-                }
-
-                if (!string.IsNullOrEmpty(localLevelId) &&
-                    !string.IsNullOrEmpty(state.LevelId) &&
-                    !string.Equals(state.LevelId, localLevelId, StringComparison.Ordinal))
-                {
-                    DisposeRemoteDownedCine(state.UserId);
-                    continue;
-                }
-
-                var client = clients[index];
-                if (client == null)
-                {
-                    // A fall/lava snapshot can dispose the remote shell because its temporary room
-                    // marker no longer matches. Recreate it immediately for the authoritative
-                    // same-level downed body instead of waiting for a sublevel transition.
-                    client = EnsureClientKingSlot(index);
-                }
-                if (client == null)
-                {
-                    DisposeRemoteDownedCine(state.UserId);
-                    continue;
-                }
-                CancelPendingClientDispose(index);
-
-                _scratchActiveCorpseIds.Add(state.UserId);
-
-                // Never create a second corpse cinematic once the local player is also downed.
-                // If this remote corpse already existed (the local player died second), keep and
-                // update that single cinematic; otherwise the brief all-down state needs no new one.
-                RemoteDownedCorpse? cine = null;
-                if (_remoteDownedCines.TryGetValue(state.UserId, out var existingCine) && existingCine != null)
-                    cine = existingCine;
-                else if (!_localFakeDead)
-                    cine = EnsureRemoteDownedCine(state, client);
-
-                if (cine != null)
-                {
-                    try
-                    {
-                        cine.UpdateTarget(
-                            state.X,
-                            state.Y,
-                            client.dir,
-                            state.HasHeadPosition ? state.HeadX : null,
-                            state.HasHeadPosition ? state.HeadY : null,
-                            state.HasHeadAnim ? state.HeadAnim : null);
-                    }
-                    catch { DisposeRemoteDownedCine(state.UserId); }
-                }
-
-                try { client._targetable = false; } catch { }
-                try { client.setPosPixel(state.X, state.Y - DownedGhostBodyYOffsetPx); } catch { }
-
-                rLastX[index] = state.X;
-                rLastY[index] = state.Y - DownedGhostBodyYOffsetPx;
-            }
-
-            if (_remoteDownedCines.Count > 0)
-            {
-                _scratchStaleCorpseIds.Clear();
-                foreach (var pair in _remoteDownedCines)
-                {
-                    if (!_scratchActiveCorpseIds.Contains(pair.Key))
-                        _scratchStaleCorpseIds.Add(pair.Key);
-                }
-
-                for (int i = 0; i < _scratchStaleCorpseIds.Count; i++)
-                    DisposeRemoteDownedCine(_scratchStaleCorpseIds[i]);
-            }
-        }
-
-        private bool IsRemoteDownedVisibleInCurrentLevel(int userId, string? localLevelId)
-        {
-            if (userId <= 0 || !_remoteDowned.TryGetValue(userId, out var state) || state == null)
-                return false;
-
-            if (string.IsNullOrWhiteSpace(localLevelId) || string.IsNullOrWhiteSpace(state.LevelId))
-                return true;
-
-            return string.Equals(localLevelId, state.LevelId, StringComparison.Ordinal);
-        }
-
-        private bool IsRemoteReviveVisibilityGraceActive(int userId)
-        {
-            if (userId <= 0 || !_remoteReviveVisibilityGraceUntilTicks.TryGetValue(userId, out var untilTicks))
-                return false;
-
-            if (Stopwatch.GetTimestamp() < untilTicks)
-                return true;
-
-            _remoteReviveVisibilityGraceUntilTicks.Remove(userId);
-            return false;
-        }
-
-        private void BeginRemoteReviveVisibilityRecovery(
-            int userId,
-            int slot,
-            GhostKing? client,
-            double x,
-            double y)
-        {
-            if (userId <= 0)
-                return;
-
-            _remoteReviveVisibilityGraceUntilTicks[userId] = Stopwatch.GetTimestamp() +
-                (long)(Stopwatch.Frequency * RemoteReviveVisibilityGraceSeconds);
-            _remoteLastDoorMarkers.Remove(userId);
-
-            if (slot < 0 || slot >= clients.Length)
-                return;
-
-            CancelPendingClientDispose(slot);
-            clientLastDownedOffsets[slot] = false;
-
-            if (client == null)
-                return;
-
-            var unusable = false;
-            try { unusable = client.destroyed; } catch { }
-            try
-            {
-                if (!unusable && me?._level != null && client._level != null &&
-                    !ReferenceEquals(client._level, me._level))
-                {
-                    unusable = true;
-                }
-            }
-            catch
-            {
-            }
-            try
-            {
-                if (!unusable && client.spr == null)
-                    unusable = true;
-            }
-            catch
-            {
-            }
-
-            if (unusable)
-            {
-                DisposeClientSlot(slot, clearIdentity: false);
-                return;
-            }
-
-            RestoreRemoteKingRenderAfterRevive(slot, client, x, y, "down-state-up");
-        }
-
-        private void RestoreRemoteKingRenderAfterRevive(
-            int slot,
-            GhostKing client,
-            double x,
-            double y,
-            string reason)
-        {
-            if (client == null || slot < 0 || slot >= clients.Length)
-                return;
-
-            try
-            {
-                if (double.IsFinite(x) && double.IsFinite(y))
-                {
-                    client.setPosPixel(x, y);
-                    rLastX[slot] = x;
-                    rLastY[slot] = y;
-                }
-            }
-            catch
-            {
-            }
-
-            var wasOutOfGame = false;
-            try { wasOutOfGame = client.isOutOfGame; } catch { }
-            try { client.lastOutOfGame = false; } catch { }
-            try { client.isOutOfGame = false; } catch { }
-            try { client.isOnScreen = true; } catch { }
-            try
-            {
-                if (client.onScreenRecent < 1200.0)
-                    client.onScreenRecent = 1200.0;
-            }
-            catch { }
-            if (wasOutOfGame)
-            {
-                try { client.onOutOfGameChange(); } catch { }
-            }
-            try { client.visible = true; } catch { }
-            try { client.spr?.set_visible(true); } catch { }
-            try { client._targetable = true; } catch { }
-
-            try { EnsureGhostKingRenderSafe(client, "remote-revive:" + reason, detachForTransition: false); } catch { }
-
-            if (clientHeads[slot] == null || client.head == null)
-                ScheduleGhostHeadRecreate(slot, immediate: true);
-            MarkGhostHeadDirty(slot, immediate: true);
-        }
-
-        private RemoteDownedCorpse? EnsureRemoteDownedCine(RemoteDownedState state, GhostKing client)
-        {
-            if (state == null || client == null || me == null)
-                return null;
-
-            if (_remoteDownedCines.TryGetValue(state.UserId, out var existing))
-            {
-                if (existing != null)
-                    return existing;
-
-                _remoteDownedCines.Remove(state.UserId);
-            }
-
-            try
-            {
-                var previousCine = dc.pr.Game.Class.ME?.curCine;
-                var created = new RemoteDownedCorpse(me, client, state.X, state.Y, client.dir, previousCine);
-                _remoteDownedCines[state.UserId] = created;
-                return created;
-            }
-            catch
-            {
-                _remoteDownedCines.Remove(state.UserId);
-                return null;
-            }
-        }
-
-        private void DisposeRemoteDownedCine(int userId)
-        {
-            if (!_remoteDownedCines.TryGetValue(userId, out var cine) || cine == null)
-                return;
-
-            _remoteDownedCines.Remove(userId);
-            try { cine.destroy(); } catch { }
-            try { cine.disposeImmediately(); } catch { }
-        }
-
-        private void DisposeAllRemoteDownedCines()
-        {
-            if (_remoteDownedCines.Count == 0)
-                return;
-
-            _scratchStaleCorpseIds.Clear();
-            foreach (var id in _remoteDownedCines.Keys)
-                _scratchStaleCorpseIds.Add(id);
-
-            for (int i = 0; i < _scratchStaleCorpseIds.Count; i++)
-                DisposeRemoteDownedCine(_scratchStaleCorpseIds[i]);
         }
 
         private bool HasAliveRemoteTeammate(NetNode net)
@@ -2026,43 +1092,6 @@ namespace DeadCellsMultiplayerMod
             _reviveHoldStartedTicks = 0;
         }
 
-        private void ShowReviveHintFor(int userId)
-        {
-            if (_remoteDownedCines.Count == 0)
-                return;
-
-            foreach (var pair in _remoteDownedCines)
-            {
-                var cine = pair.Value;
-                if (cine == null)
-                    continue;
-
-                try
-                {
-                    if (pair.Key == userId)
-                        cine.SetInteractionLabel(Localize(ReviveHintText));
-                    else
-                        cine.SetInteractionLabel(null);
-                }
-                catch
-                {
-                }
-            }
-        }
-
-        private void ClearReviveHints()
-        {
-            if (_remoteDownedCines.Count == 0)
-                return;
-
-            foreach (var cine in _remoteDownedCines.Values)
-            {
-                if (cine == null)
-                    continue;
-                try { cine.SetInteractionLabel(null); } catch { }
-            }
-        }
-
         private bool TryConsumeOneFlask(Hero hero)
         {
             if (hero == null)
@@ -2156,35 +1185,6 @@ namespace DeadCellsMultiplayerMod
                 return levelId.Trim();
 
             return string.Empty;
-        }
-
-        private void StartLocalDeadCine(Hero hero)
-        {
-            if (hero == null)
-                return;
-
-            if (_localDeadCine != null)
-                return;
-
-            try
-            {
-                _localDeadCine = new DeadBase(hero, ModEntry.GetPrimaryClient());
-            }
-            catch
-            {
-                _localDeadCine = null;
-            }
-        }
-
-        private void StopLocalDeadCine()
-        {
-            var cine = _localDeadCine;
-            _localDeadCine = null;
-            if (cine == null)
-                return;
-
-            try { cine.destroy(); } catch { }
-            try { cine.disposeImmediately(); } catch { }
         }
 
         private void ResetFakeDeathState(
@@ -2365,38 +1365,6 @@ namespace DeadCellsMultiplayerMod
             _allDownedRestartAtTicks = 0;
         }
 
-        private bool TryUpdateDownedPositionFromCorpse(double corpseX, double corpseY)
-        {
-            // Co-op corpses are pinned to the authoritative revive point. Never let a one-frame
-            // corpse physics step move the gameplay anchor downward or through floor tiles.
-            if (_localFakeDead && ShouldAnchorLocalDownedCorpse())
-                return false;
-
-            if (!double.IsFinite(corpseX) || !double.IsFinite(corpseY))
-                return false;
-
-            if (!_hasLocalDownedAnchor)
-            {
-                _localDownedAnchorX = _localDownedX;
-                _localDownedAnchorY = _localDownedY;
-                _hasLocalDownedAnchor = true;
-            }
-
-            var dx = corpseX - _localDownedAnchorX;
-            var dy = corpseY - _localDownedAnchorY;
-            var distSq = dx * dx + dy * dy;
-            if (distSq > DownedCorpseMaxDriftSq)
-                return false;
-
-            _localDownedX = corpseX;
-            _localDownedY = corpseY;
-            _localHeldX = _localDownedX;
-            _localHeldY = _localDownedY;
-            _localDownedAnchorX = corpseX;
-            _localDownedAnchorY = corpseY;
-            return true;
-        }
-
         private static void EnsureHeroVisibilityAfterRoomChange(Hero? hero)
         {
             if (hero == null)
@@ -2448,3 +1416,4 @@ namespace DeadCellsMultiplayerMod
         }
     }
 }
+
