+++ README_ru.md
@@ -106,8 +106,10 @@
 - [x] Синхронизация загрузки графа уровня (босс-клетки, переходы)  
 - [x] Слоты сохранения мультиплеера и продолжение  
 - [x] Камера-спектатор  
-- [ ] Кастомный режим  
+- [x] Кастомный режим  
 - [x] Steam P2P подключение  
+
+**Примечание:** синхронизация врагов использует host-owned NetId (не нативные id игры). `AdvancedCoop` — только lobby heartbeat и постоянный прогресс, не синкает врагов.
 
 ---
